add a bunch of stuff, found technical debt limit
At the moment the current setup of data, datatypes, extension methods, etc are not fit for purpose and a serious overhaul is needed for the app to truly work well. This branch will stop here and I will start an overhaul branch.

+++ BackEnd/WishGranter/ExtensionMethods/WG_Calculation.cs
@@ -343,7 +343,7 @@ namespace WishGranterProto.ExtensionMethods
 
                     // Package details in Transmission object
                     TransmissionArmorTestShot testShot = new();
-                    testShot.DurabilityPerc = (startingDura - doneDamage) / armorItem.MaxDurability * 100;
+                    testShot.DurabilityPerc = (double)((startingDura - doneDamage) / armorItem.MaxDurability * 100);
                     testShot.DoneDamage = doneDamage;
                     testShot.Durability = startingDura - doneDamage;
                     testShot.PenetrationChance = penetrationChance;
@@ -479,7 +479,7 @@ namespace WishGranterProto.ExtensionMethods
 
                     // Package details in Transmission object
                     TransmissionArmorTestShot testShot = new();
-                    testShot.DurabilityPerc = (startingDura - doneDamage) / armorItem.MaxDurability * 100;
+                    testShot.DurabilityPerc = (double)((startingDura - doneDamage) / armorItem.MaxDurability * 100);
                     testShot.DoneDamage = doneDamage;
                     testShot.Durability = startingDura - doneDamage;
                     testShot.PenetrationChance = penetrationChance;
